[WPE] Implement ImageAdapter https://bugs.webkit.org/show_bug.cgi?id=208869
Reviewed by Carlos Garcia Campos.

Until now WPE didn't have a specific implementation of ImageAdapter,
and used a generic one instead.  This affected the results of some
layout tests.

It turns out that some parts of ImageAdapterGtk can be actually
reused for WPE, particularly the code that loads bundled images from
a gresource file.

Also, modify PlatformWPE.cmake to bundle images into gresource file.

* LayoutTests/platform/glib/editing/execCommand/insertImage-expected.txt: Renamed from LayoutTests/platform/gtk/editing/execCommand/insertImage-expected.txt.
* LayoutTests/platform/glib/fast/borders/rtl-border-05-expected.txt: Renamed from LayoutTests/platform/gtk/fast/borders/rtl-border-05-expected.txt.
* LayoutTests/platform/glib/fast/dom/34176-expected.txt: Renamed from LayoutTests/platform/gtk/fast/dom/34176-expected.txt.
* LayoutTests/platform/glib/fast/dom/inner-text-expected.txt: Renamed from LayoutTests/platform/gtk/fast/dom/inner-text-expected.txt.
* LayoutTests/platform/glib/fast/encoding/utf-16-big-endian-expected.txt: Renamed from LayoutTests/platform/gtk/fast/encoding/utf-16-big-endian-expected.txt.
* LayoutTests/platform/glib/fast/encoding/utf-16-little-endian-expected.txt: Renamed from LayoutTests/platform/gtk/fast/encoding/utf-16-little-endian-expected.txt.
* LayoutTests/platform/glib/fast/hidpi/broken-image-icon-hidpi-expected.txt: Renamed from LayoutTests/platform/gtk/fast/hidpi/broken-image-icon-hidpi-expected.txt.
* LayoutTests/platform/glib/fast/hidpi/broken-image-icon-very-hidpi-expected.txt: Renamed from LayoutTests/platform/gtk/fast/hidpi/broken-image-icon-very-hidpi-expected.txt.
* LayoutTests/platform/glib/tables/mozilla/bugs/bug2962-expected.txt: Renamed from LayoutTests/platform/gtk/tables/mozilla/bugs/bug2962-expected.txt.
* LayoutTests/platform/wpe/TestExpectations:
* LayoutTests/platform/wpe/editing/execCommand/insertImage-expected.txt:
* LayoutTests/platform/wpe/fast/borders/rtl-border-05-expected.txt:
* LayoutTests/platform/wpe/fast/dom/34176-expected.txt:
* LayoutTests/platform/wpe/fast/dom/inner-text-expected.txt:
* LayoutTests/platform/wpe/fast/encoding/utf-16-big-endian-expected.txt:
* LayoutTests/platform/wpe/fast/encoding/utf-16-little-endian-expected.txt:
* LayoutTests/platform/wpe/fast/hidpi/broken-image-icon-hidpi-expected.txt:
* LayoutTests/platform/wpe/fast/hidpi/broken-image-icon-very-hidpi-expected.txt:
* LayoutTests/platform/wpe/tables/mozilla/bugs/bug2962-expected.txt:
* LayoutTests/platform/wpe/tables/mozilla/bugs/bug56201-expected.txt: Removed.
* Source/WebCore/SourcesGTK.txt:
* Source/WebCore/SourcesWPE.txt:
* Source/WebCore/platform/graphics/ImageAdapter.cpp:
* Source/WebCore/platform/graphics/glib/ImageAdapterGLib.cpp: Copied from Source/WebCore/platform/graphics/gtk/ImageAdapterGtk.cpp.
(WebCore::loadImageFromGResource):
(WebCore::ImageAdapter::loadPlatformResource):
(WebCore::ImageAdapter::invalidate):
* Source/WebCore/platform/graphics/gtk/ImageAdapterGtk.cpp:
(WebCore::loadImageFromGResource): Deleted.
(WebCore::ImageAdapter::loadPlatformResource): Deleted.
(WebCore::ImageAdapter::invalidate): Deleted.
* Source/WebKit/PlatformGTK.cmake:
* Source/WebKit/PlatformWPE.cmake:
* Source/WebKit/Resources/wpe/wpe-theme.css: Added.
(webkitwebview > dimming):
(webkitwebview > shadow):
(webkitwebview > shadow.right):
(webkitwebview > shadow.left):
(webkitwebview > border):
(webkitwebview > outline):

Canonical link: https://commits.webkit.org/299620@main
