+++ LayoutTests/platform/glib/fast/borders/rtl-border-05-expected.txt
