+++ LayoutTests/platform/glib/fast/dom/inner-text-expected.txt
