+++ LayoutTests/platform/glib/fast/encoding/utf-16-big-endian-expected.txt
