+++ LayoutTests/platform/glib/fast/encoding/utf-16-little-endian-expected.txt
