+++ LayoutTests/platform/glib/fast/hidpi/broken-image-icon-hidpi-expected.txt
