+++ LayoutTests/platform/glib/fast/hidpi/broken-image-icon-very-hidpi-expected.txt
