+++ LayoutTests/platform/glib/tables/mozilla/bugs/bug2962-expected.txt
