+++ LayoutTests/platform/wpe/TestExpectations
@@ -107,8 +107,6 @@ pointerevents/mouse [ Pass ]
 # Test passing for WPE but failing in GTK.
 imported/w3c/web-platform-tests/css/css-transforms/transform3d-preserve3d-013.html [ Pass ]
 imported/w3c/web-platform-tests/css/css-text/hyphens/hyphens-auto-control.html [ Pass ]
-imported/w3c/web-platform-tests/css/css-ui/resize-child-will-change-transform.html [ Pass ]
-imported/w3c/web-platform-tests/css/css-ui/resize-generated-content.html [ Pass ]
 
 http/tests/cookies/same-site [ Pass ]
 
@@ -384,9 +382,6 @@ webkit.org/b/181767 fast/hidpi/filters-hue-rotate.html [ Pass ImageOnlyFailure ]
 webkit.org/b/206653 imported/w3c/web-platform-tests/clipboard-apis/async-navigator-clipboard-basics.https.html [ Failure Pass ]
 
 webkit.org/b/285785 printing/background-clip-text.html [ ImageOnlyFailure ]
-
-webkit.org/b/208869 fast/images/inline-image-box-with-no-alt-should-collapse-no-quirks.html [ Failure ]
-webkit.org/b/208869 fast/images/inline-image-box-with-no-alt-should-collapse-quirks.html [ Failure ]
 
 webkit.org/b/208985 fast/viewport/ios/shrink-to-fit-large-content-width.html [ Failure ]
