--- LayoutTests/platform/wpe/fast/borders/rtl-border-05-expected.txt
@@ -1,55 +1,0 @@
-layer at (0,0) size 800x600
-  RenderView at (0,0) size 800x600
-layer at (0,0) size 800x582
-  RenderBlock {HTML} at (0,0) size 800x582
-    RenderBody {BODY} at (8,16) size 784x558
-      RenderBlock {P} at (0,0) size 784x38
-        RenderText {#text} at (0,1) size 760x36
-          text run at (0,1) width 760: "Test that when we have images sprinkled within a span, borders, paddings and margins are still"
-          text run at (0,20) width 168: "calculated correctly ."
-      RenderBlock {DIV} at (0,54) size 212x252 [border: (1px solid #000000)]
-        RenderBlock {P} at (6,22) size 200x160
-          RenderText {#text} at (0,7) size 189x49
-            text run at (0,7) width 189: "aaa bbb ccc ddd eee fff"
-            text run at (0,39) width 105 RTL: "\x{5DC}\x{5DE}\x{5D4} \x{5E2}\x{5E9}\x{5D9}\x{5EA}\x{5D9} \x{5D6}\x{5D0}\x{5EA}"
-            text run at (104,39) width 6: " "
-            text run at (109,39) width 21: "gg"
-          RenderInline {SPAN} at (0,0) size 200x89 [border: (1px solid #0000FF) (10px solid #FFA500) (1px solid #0000FF) (10px solid #008000)]
-            RenderText {#text} at (0,71) size 48x17
-              text run at (0,71) width 48: "dgnjk "
-            RenderImage {IMG} at (48,79) size 6x6 [border: (1px solid #0000FF)]
-            RenderText {#text} at (54,71) size 48x17
-              text run at (54,71) width 48: "njkng "
-          RenderText {#text} at (0,71) size 189x81
-            text run at (147,71) width 39 RTL: "\x{5D6}\x{5D0}\x{5EA} \x{2013}"
-            text run at (0,103) width 99 RTL: "\x{5D0}\x{5D9}\x{5E0}\x{5D9} \x{5D9}\x{5D5}\x{5D3}\x{5E2} \x{5D1}\x{5E2}\x{5E6}\x{5DE}"
-            text run at (98,103) width 91: " qqq rrr sss"
-            text run at (0,135) width 85: "ttt uuu vvv"
-        RenderBlock {P} at (6,198) size 200x32
-          RenderInline {SPAN} at (25,6) size 164x19 [border: (1px solid #0000FF) (10px solid #FFA500) (1px solid #0000FF) (10px solid #008000)]
-            RenderText {#text} at (45,7) size 105x17
-              text run at (45,7) width 105: "Some fill text"
-            RenderImage {IMG} at (149,17) size 5x4
-      RenderBlock {DIV} at (0,306) size 212x252 [border: (1px solid #000000)]
-        RenderBlock {P} at (6,22) size 200x160
-          RenderText {#text} at (11,7) size 189x49
-            text run at (11,7) width 189: "aaa bbb ccc ddd eee fff"
-            text run at (25,39) width 21: "gg"
-            text run at (45,39) width 110 RTL: "\x{5DC}\x{5DE}\x{5D4} \x{5E2}\x{5E9}\x{5D9}\x{5EA}\x{5D9} \x{5D6}\x{5D0}\x{5EA} "
-          RenderInline {SPAN} at (78,0) size 122x89 [border: (1px solid #0000FF) (10px solid #FFA500) (1px solid #0000FF) (10px solid #008000)]
-            RenderText {#text} at (103,71) size 48x17
-              text run at (103,71) width 48: "dgnjk "
-            RenderImage {IMG} at (151,79) size 6x6 [border: (1px solid #0000FF)]
-            RenderText {#text} at (98,71) size 102x17
-              text run at (98,71) width 5 RTL: " "
-              text run at (157,71) width 43: "njkng"
-          RenderText {#text} at (11,71) size 189x81
-            text run at (14,71) width 39 RTL: "\x{5D6}\x{5D0}\x{5EA} \x{2013}"
-            text run at (11,103) width 86: "qqq rrr sss"
-            text run at (96,103) width 104 RTL: "\x{5D0}\x{5D9}\x{5E0}\x{5D9} \x{5D9}\x{5D5}\x{5D3}\x{5E2} \x{5D1}\x{5E2}\x{5E6}\x{5DE} "
-            text run at (115,135) width 85: "ttt uuu vvv"
-        RenderBlock {P} at (6,198) size 200x32
-          RenderInline {SPAN} at (26,6) size 164x19 [border: (1px solid #0000FF) (10px solid #FFA500) (1px solid #0000FF) (10px solid #008000)]
-            RenderText {#text} at (50,7) size 105x17
-              text run at (50,7) width 105: "Some fill text"
-            RenderImage {IMG} at (46,17) size 5x4
