--- LayoutTests/platform/wpe/fast/dom/inner-text-expected.txt
@@ -1,9 +1,0 @@
-layer at (0,0) size 800x600
-  RenderView at (0,0) size 800x600
-layer at (0,0) size 800x600
-  RenderBlock {HTML} at (0,0) size 800x600
-    RenderBody {BODY} at (8,8) size 784x584
-      RenderImage {IMG} at (0,10) size 4x4
-      RenderText {#text} at (4,0) size 221x17
-        text run at (4,0) width 221: " This text should only appear once."
-      RenderText {#text} at (0,0) size 0x0
