--- LayoutTests/platform/wpe/fast/hidpi/broken-image-icon-hidpi-expected.txt
@@ -1,11 +1,0 @@
-layer at (0,0) size 800x600
-  RenderView at (0,0) size 800x600
-layer at (0,0) size 800x600
-  RenderBlock {HTML} at (0,0) size 800x600
-    RenderBody {BODY} at (8,8) size 784x584
-      RenderBlock {DIV} at (0,0) size 784x18
-        RenderText {#text} at (0,0) size 738x17
-          text run at (0,0) width 738: "This test passes if the broken image icon appears high-resolution in HiDPI and if the outline is of the appropriate size."
-      RenderBlock (anonymous) at (0,18) size 784x4
-        RenderImage {IMG} at (0,0) size 4x4
-        RenderText {#text} at (0,0) size 0x0
