--- LayoutTests/platform/wpe/tables/mozilla/bugs/bug2962-expected.txt
@@ -1,22 +1,0 @@
-layer at (0,0) size 800x600
-  RenderView at (0,0) size 800x600
-layer at (0,0) size 800x600
-  RenderBlock {HTML} at (0,0) size 800x600
-    RenderBody {BODY} at (8,8) size 784x584
-      RenderTable {TABLE} at (0,0) size 300x46 [border: (2px outset #000000)]
-        RenderTableSection {TBODY} at (2,2) size 296x42
-          RenderTableRow {TR} at (0,0) size 296x42
-            RenderTableCell {TD} at (0,0) size 296x42 [bgcolor=#FFFF00] [border: (1px inset #000000)] [r=0 c=0 rs=1 cs=1]
-              RenderBlock {CENTER} at (3,3) size 290x36
-                RenderInline {A} at (56,0) size 0x17 [color=#0000EE]
-                  RenderImage {IMG} at (0,0) size 4x4
-                RenderText {#text} at (0,0) size 0x0
-                RenderInline {A} at (56,0) size 185x35 [color=#0000EE]
-                  RenderInline {FONT} at (56,0) size 185x35 [color=#000000]
-                    RenderText {#text} at (56,0) size 185x17
-                      text run at (56,0) width 111: "Free Homepage! "
-                      text run at (166,0) width 75: "Free Email!"
-                    RenderBR {BR} at (240,0) size 1x17
-                    RenderText {#text} at (60,18) size 170x17
-                      text run at (60,18) width 41: "It's all "
-                      text run at (101,18) width 129: "yours at ACMEcity!"
