--- LayoutTests/platform/wpe/tables/mozilla/bugs/bug56201-expected.txt
@@ -1,18 +1,0 @@
-layer at (0,0) size 800x600
-  RenderView at (0,0) size 800x600
-layer at (0,0) size 800x54
-  RenderBlock {HTML} at (0,0) size 800x54
-    RenderBody {BODY} at (8,8) size 784x38
-      RenderTable {TABLE} at (0,0) size 34x38 [border: (1px outset #000000)]
-        RenderTableSection {TBODY} at (1,1) size 32x36
-          RenderTableRow {TR} at (0,2) size 32x32
-            RenderTableCell {TD} at (2,2) size 28x32 [border: (1px inset #000000)] [r=0 c=0 rs=1 cs=1]
-              RenderTable {TABLE} at (2,2) size 24x28 [border: (1px outset #000000)]
-                RenderTableSection {TBODY} at (1,1) size 22x26
-                  RenderTableRow {TR} at (0,2) size 22x22
-                    RenderTableCell {TD} at (2,2) size 8x22 [border: (1px inset #000000)] [r=0 c=0 rs=1 cs=1]
-                      RenderImage {IMG} at (2,12) size 4x4
-                      RenderText {#text} at (0,0) size 0x0
-                    RenderTableCell {TD} at (12,2) size 8x22 [border: (1px inset #000000)] [r=0 c=1 rs=1 cs=1]
-                      RenderImage {IMG} at (2,12) size 4x4
-                      RenderText {#text} at (0,0) size 0x0
