+++ Source/WebCore/SourcesGTK.txt
@@ -79,6 +79,8 @@ platform/graphics/gbm/GraphicsContextGLTextureMapperGBM.cpp @no-unify
 platform/graphics/gbm/MemoryMappedGPUBuffer.cpp
 platform/graphics/gbm/PlatformDisplayGBM.cpp @no-unify
 
+platform/graphics/glib/ImageAdapterGLib.cpp
+
 platform/graphics/gtk/ColorGtk.cpp
 platform/graphics/gtk/GdkCairoUtilities.cpp
 platform/graphics/gtk/GdkSkiaUtilities.cpp
