+++ Source/WebCore/platform/graphics/gtk/ImageAdapterGtk.cpp
@@ -29,32 +29,12 @@
 #include "BitmapImage.h"
 #include "GdkCairoUtilities.h"
 #include "GdkSkiaUtilities.h"
-#include "SharedBuffer.h"
 #include <cairo.h>
 #include <gdk/gdk.h>
 #include <wtf/glib/GSpanExtras.h>
 #include <wtf/glib/GUniquePtr.h>
 
 namespace WebCore {
-
-static Ref<Image> loadImageFromGResource(const char* iconName)
-{
-    auto icon = BitmapImage::create();
-    GUniquePtr<char> path(g_strdup_printf("/org/webkitgtk/resources/images/%s", iconName));
-    GRefPtr<GBytes> data = adoptGRef(g_resources_lookup_data(path.get(), G_RESOURCE_LOOKUP_FLAGS_NONE, nullptr));
-    ASSERT(data);
-    icon->setData(SharedBuffer::create(span(data)), true);
-    return icon;
-}
-
-Ref<Image> ImageAdapter::loadPlatformResource(const char* name)
-{
-    return loadImageFromGResource(name);
-}
-
-void ImageAdapter::invalidate()
-{
-}
 
 GRefPtr<GdkPixbuf> ImageAdapter::gdkPixbuf()
 {
